refactor: Cleanup `remove` tests
